+++ app/src/main/java/eu/kanade/tachiyomi/data/database/models/TrackImpl.kt
@@ -18,7 +18,7 @@ class TrackImpl : Track {
 
     override var total_chapters: Long = 0L
 
-    override var score: Float = 0f
+    override var score: Double = 0.0
 
     override var status: Int = 0
