+++ app/src/main/java/eu/kanade/tachiyomi/data/track/bangumi/dto/BGMSearch.kt
@@ -33,7 +33,7 @@ data class BGMSearchItem(
         } else {
             this@BGMSearchItem.summary.orEmpty()
         }
-        score = rating?.score?.toFloat() ?: -1.0F
+        score = rating?.score?.toDouble() ?: -1.0
         tracking_url = url
         total_chapters = epsCount ?: 0
     }
